Move plan Merge methods to their own file for easier diffing
