+++ internal/plan/plan.go
@@ -90,63 +90,6 @@ type Service struct {
 	BackoffLimit   OptionalDuration         `yaml:"backoff-limit,omitempty"`
 }
 
-// Merge merges the fields set in other into c.
-func (s *Service) Merge(other *Service) {
-	if other.Summary != "" {
-		s.Summary = other.Summary
-	}
-	if other.Description != "" {
-		s.Description = other.Description
-	}
-	if other.Startup != StartupUnknown {
-		s.Startup = other.Startup
-	}
-	if other.Command != "" {
-		s.Command = other.Command
-	}
-	if other.UserID != nil {
-		s.UserID = other.UserID
-	}
-	if other.User != "" {
-		s.User = other.User
-	}
-	if other.GroupID != nil {
-		s.GroupID = other.GroupID
-	}
-	if other.Group != "" {
-		s.Group = other.Group
-	}
-	s.Before = append(s.Before, other.Before...)
-	s.After = append(s.After, other.After...)
-	for k, v := range other.Environment {
-		if s.Environment == nil {
-			s.Environment = make(map[string]string)
-		}
-		s.Environment[k] = v
-	}
-	if other.OnSuccess != "" {
-		s.OnSuccess = other.OnSuccess
-	}
-	if other.OnFailure != "" {
-		s.OnFailure = other.OnFailure
-	}
-	for k, v := range other.OnCheckFailure {
-		if s.OnCheckFailure == nil {
-			s.OnCheckFailure = make(map[string]ServiceAction)
-		}
-		s.OnCheckFailure[k] = v
-	}
-	if other.BackoffDelay.IsSet {
-		s.BackoffDelay = other.BackoffDelay
-	}
-	if other.BackoffFactor.IsSet {
-		s.BackoffFactor = other.BackoffFactor
-	}
-	if other.BackoffLimit.IsSet {
-		s.BackoffLimit = other.BackoffLimit
-	}
-}
-
 // Equal returns true when the two services are equal in value.
 func (s *Service) Equal(other *Service) bool {
 	if s == other {
@@ -199,40 +142,6 @@ type Check struct {
 	Exec *ExecCheck `yaml:"exec,omitempty"`
 }
 
-// Merge merges the fields set in other into c.
-func (c *Check) Merge(other *Check) {
-	if other.Level != "" {
-		c.Level = other.Level
-	}
-	if other.Period.IsSet {
-		c.Period = other.Period
-	}
-	if other.Timeout.IsSet {
-		c.Timeout = other.Timeout
-	}
-	if other.Failures != 0 {
-		c.Failures = other.Failures
-	}
-	if other.HTTP != nil {
-		if c.HTTP == nil {
-			c.HTTP = &HTTPCheck{}
-		}
-		c.HTTP.Merge(other.HTTP)
-	}
-	if other.TCP != nil {
-		if c.TCP == nil {
-			c.TCP = &TCPCheck{}
-		}
-		c.TCP.Merge(other.TCP)
-	}
-	if other.Exec != nil {
-		if c.Exec == nil {
-			c.Exec = &ExecCheck{}
-		}
-		c.Exec.Merge(other.Exec)
-	}
-}
-
 // CheckLevel specifies the optional check level.
 type CheckLevel string
 
@@ -248,33 +157,10 @@ type HTTPCheck struct {
 	Headers map[string]string `yaml:"headers,omitempty"`
 }
 
-// Merge merges the fields set in other into c.
-func (c *HTTPCheck) Merge(other *HTTPCheck) {
-	if other.URL != "" {
-		c.URL = other.URL
-	}
-	for k, v := range other.Headers {
-		if c.Headers == nil {
-			c.Headers = make(map[string]string)
-		}
-		c.Headers[k] = v
-	}
-}
-
 // TCPCheck holds the configuration for an HTTP health check.
 type TCPCheck struct {
 	Port int    `yaml:"port,omitempty"`
 	Host string `yaml:"host,omitempty"`
-}
-
-// Merge merges the fields set in other into c.
-func (c *TCPCheck) Merge(other *TCPCheck) {
-	if other.Port != 0 {
-		c.Port = other.Port
-	}
-	if other.Host != "" {
-		c.Host = other.Host
-	}
 }
 
 // ExecCheck holds the configuration for an exec health check.
@@ -286,36 +172,6 @@ type ExecCheck struct {
 	GroupID     *int              `yaml:"group-id,omitempty"`
 	Group       string            `yaml:"group,omitempty"`
 	WorkingDir  string            `yaml:"working-dir,omitempty"`
-}
-
-// Merge merges the fields set in other into c.
-func (c *ExecCheck) Merge(other *ExecCheck) {
-	if other.Command != "" {
-		c.Command = other.Command
-	}
-	for k, v := range other.Environment {
-		if c.Environment == nil {
-			c.Environment = make(map[string]string)
-		}
-		c.Environment[k] = v
-	}
-	if other.UserID != nil {
-		userID := *other.UserID
-		c.UserID = &userID
-	}
-	if other.User != "" {
-		c.User = other.User
-	}
-	if other.GroupID != nil {
-		groupID := *other.GroupID
-		c.GroupID = &groupID
-	}
-	if other.Group != "" {
-		c.Group = other.Group
-	}
-	if other.WorkingDir != "" {
-		c.WorkingDir = other.WorkingDir
-	}
 }
 
 // FormatError is the error returned when a layer has a format error, such as
